better check for missing
Former-commit-id: aba85a4061ae2d1d68a51373d4f01452777c37d1

+++ openproblems/data/__init__.py
@@ -17,10 +17,11 @@ TEMPDIR = _make_tempdir()
 
 
 def _cleanup():
-    try:
-        shutil.rmtree(TEMPDIR)
-    except (FileNotFoundError, PermissionError):
-        pass
+    if os.path.isdir(TEMPDIR):
+        try:
+            shutil.rmtree(TEMPDIR)
+        except (FileNotFoundError, PermissionError):
+            pass
 
 
 atexit.register(_cleanup)
